Added students.txt which has student click-thru flow
To this text file I added a flow of what a student should see when they click on the log in button and internally also what needs to happen. I describe the properties of student "objects" that will be stored in this database as well. Importantly, we will keep track of the classes they have taken as well as the minors we have recommended for them before and the progress they have made so far.