Created Student DB and student DB retriever
The two programs I created do the following:

create_student_database creates the collection of students and assigns the indices associated with students: netid, name, major, classes, and recommendations

student_database.py: This is what we would want to use to display class information on the class board! Basically, if a student doesn't exist in the database, the student gets added to the database through this program. If they are already on the database, then you proceed as usual. In addition, you can get the classes of a given student by passing in their username into a get_student_classes method. You can update the classes of a student similarly by passing in the classes to be updated and the student's username as well.