Returning User Functionality
Changes are as follows:
application.py: Refactored the routing system. Now when someone logs in, check to see if they are logging in for the first time and send them to loginpage.html. If they are a returning user, reroute them to \classboard. The other major change is storing the username in session when someone logs in. This way, we can access the username anywhere in application.py by 'getting' the username from the flask.session. In addition, I added an additional routing /profile so that when a user wants to change their name and major, they can do so and the information will be updated on the backend

student_database.py: the main new thing I added here was a function that updates the user's name and major in the backend. This is done through "update_student_profile"

loginpage.html: On this page, I just modified the routing to go to \profile instead of \classboard

figma_classboard.html: I added a div element before "Your Class Board" span element for readability purposes. I also added two <br> after the "Your Class Board" element so that there is more space between that statement and the ones below for readability.